docs,refactor(indexer-nibi): functions docs and code cleanup
