+++ packages/indexer-nibi/src/types.ts
@@ -1,5 +1,3 @@
-import { CandlePeriod } from "./constant"
-
 export interface TypeMarkPrice {
   pair: string
   price: string
